requirements.txt
Flask==2.3.3
google-cloud-bigquery==3.13.0
google-auth==2.23.4
gunicorn==21.2.0 